+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixes
 
-- Fix `event.origin` and `event.environment` on unhandled exception on iOS ([#3041](https://github.com/getsentry/sentry-react-native/pull/3041))
+- Fix `event.origin` and `event.environment` on unhandled exception ([#3041](https://github.com/getsentry/sentry-react-native/pull/3041))
 
 ## 5.4.1