+++ CHANGELOG.md
@@ -7,6 +7,12 @@
 - Fix `event.origin` and `event.environment` on unhandled exception ([#3041](https://github.com/getsentry/sentry-react-native/pull/3041))
 - Don't pass `enableTracing` from RN to `sentry-cocoa` options ([#3042](https://github.com/getsentry/sentry-react-native/pull/3042))
 
+### Dependencies
+
+- Bump JavaScript SDK from v7.50.0 to v7.51.0 ([#3043](https://github.com/getsentry/sentry-react-native/pull/3043))
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#7510)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/7.50.0...7.51.0)
+
 ## 5.4.1
 
 ### Fixes