+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix `event.origin` and `event.environment` on unhandled exception ([#3041](https://github.com/getsentry/sentry-react-native/pull/3041))
 - Don't pass `enableTracing` from RN to `sentry-cocoa` options ([#3042](https://github.com/getsentry/sentry-react-native/pull/3042))
+- Only store envelops of fatal crashes on iOS ([#3051](https://github.com/getsentry/sentry-react-native/pull/3051))
 
 ### Dependencies